show readme preview on repository files view
When entering a repository, a file listing is showed as well as a README
preview, similar to the view of major git hosting services, like Github,
Gitlab, Codeberg, etc.

When browsing through the repository files, the README preview is not
showed, again, imitating the behaviour of more popular git hosting
services.